Improve GUI integration plus timing/debugging/clean up
Boss 'render_all' function now calls an optional application provided
GUI draw function. This is in the form of a 'GuiTrait' which is
implemented for the existing eGui integration but could also be
an application provided implementation.

Timing/debugging of the 'render_all' loop. The wait on acquire future
that was added as a Nvidia + Linux + vsync work around was modified
to only be used when vsync is on.

Some Boss fields were made private with accessors added. This is a
breaking change for some applications.