Many updates
Change all shader names back to .vert and .frag since these are the default
extensions used by the shader compiler.

Build main render pass shaders by calling the shader compiler directly in
a build.rs file instead of using the vulkano shader macro.

Use shared shader layouts and self defined structs for the main render pass
shaders instead of relying on reflection from the vulkano shader macro
(since it is no longer used).

Use push constants with separate ranges for the main render pass vertex and
fragment shaders since the vulkano shader macro will no longer incorrectly
reject these (since it is no longer used).

Pass the projection matrix and an array of model-view matrices to the main
render pass shaders in a SSBO. The array does not have a fixed size. A
matrix id is sent in a push constant so that the shader can access the
correct array element.

Move lots of main render pass functionality from the pipeline submodule to
the "layout" submodule, or to a "conduit" structure which combines the
pipeline with the device vertex buffers, or to the model manager module
itself.

Draw first iterates through all models. For those marked as visible, the
model-view matrix is calculated and stored so that it can be added to the
SSBO later. The matrix index is added to a hashmap. Drawing is done on a
per conduit basis with models only drawn if they are in the hashmap.