Change validation callback to avoid panic with Rust 1.78
The vulkano 0.34.1 code for a callback from validation layers does not
handle null pointers correctly. Previously Rust could not detect this
and things appeared to work correctly. Rust updates in version 1.78
allow it to detect the null pointer and panic at startup. The issue
has been fixed in vulkano but not released.

This project was not using most of the data anyway so it has been changed
to use a simpler validation callback that avoids the issue.